refine e2e test
